feat: Add functions to initialize blockchain and server
This commit adds two new functions, `initialize_blockchain` and `initialize_server`, to the `global.rs` file. The `initialize_blockchain` function initializes the blockchain by cloning the `CHAIN` and returning it as an `Arc<Mutex<Chain>>`. The `initialize_server` function initializes the server by cloning the `CHAIN` and `POOL` and creating a new `P2p` instance. These functions are used in the `main.rs` file to set up the blockchain and server before starting them.